more code reorganization
