+++ impl/src/code_gen/read/mod.rs
@@ -1,3 +1,4 @@
+pub mod helpers;
 pub mod parsely_read_enum_data;
 pub mod parsely_read_field_data;
 pub mod parsely_read_struct_data;
