--- impl/src/parsely_data/mod.rs
@@ -1,2 +1,0 @@
-pub mod parsely_common_field_data;
-pub mod read;
